make the `-near` crates publishable
We have a bunch of near forks of wasmer 0.17 published

* https://lib.rs/crates/wasmer-runtime-core-near
* https://lib.rs/crates/wasmer-runtime-near
* https://lib.rs/crates/wasmer-singlepass-backend-near

Turns out we don't have the exact code for those published, this commit
remedies situation.

The 0.17.2 branch contains almost the right code -- it has all the
patches, and misses just the literal renames. This commit adds the
renames. Unlike the published code, the renames are done via  `package`
key in Cargo.toml rather than via renaming rs files.

I diffed the code with the results of `cargo vendor` in nearcore to make
sure that there are no unexpected differences.