fix(wapm) Fix the `wax` script generation.
On macOS, `echo` always enables interpretation of backslach sequences
(like `\n`). On Linux, the `-e` flag must be used. Even if this `-e`
flag is absent from the macOS man's page for `echo(1)`, it is accepted
and ignored, thus we can safely use it in our `Makefile`.

To test this patch:

```sh
$ # before
$ make package-wapm
$ cat package/bin/wax
$
$ # after
$ make package-wapm
$ cat package/bin/wax
wapm execute "$@"
```