Remove FuncRef registry
This backports
  https://github.com/wasmerio/wasmer/pull/2437
to fix
  https://github.com/wasmerio/wasmer/issues/2638

On a high-level, Store (global state for the whole wasmer runtime) used
to store pointers to host functions. This was a partial implementation
of the module linking proposal: for that, we need to be able to use
functions *across* several instances, so we need to keep such functions
in some sort of global state.

This commit moves the data to the Instance. This breaks some module
linking related tests, but we are not using that, and the support wasn't
complete anyway.

Note that the actual memory management remains the same:
`Instance::imported_function_envs` is the thing that was and still is
responsible for dropping data related to external functions.