doc(makefile) Make `sed -i` cross-platform to fix `wasmer-c-api`'s doc.
When generating the documentation for `wasmer-c-api`, we need to
modify its `Cargo.toml` file. For that, we use `sed`. But we must use
the `-i` argument (for `in-place` modifications). Unfortunately, it's
different between macOS and Linux. This patch makes it cross-platform
so that it runs on macOS and on Linux (incl. the CI).