fix(vfs) Update `#[cfg(not(unix))]` to `#[cfg(windows)]`.
The `cfg(unix)` or `cfg(windows)` syntax are shortcuts to
`cfg(target_family = "unix")` and `cfg(target_family = "windows")`.

If we check the list of target families,
https://doc.rust-lang.org/reference/conditional-compilation.html#target_family,
we have:

* `unix`,
* `windows`,
* `wasm`.

So by saying `not(unix)`, we may say `windows`, but also `wasm`. In
this case, we specifically want `windows`. So let's change `not(unix)`
to `windows`, which is clearer and more correct.