feat(vfs) Unified API.
Unfortunately, this patch contains multiple features.

First off, it renames the types from `host_fs` and `mem_fs` so that
they use the same names. Apart from the module name, the types owned
by the module are the same, which simplifies the usage greatly.

Second, the code has been cleaned up a little bit. The more important
clean up is probably the renaming of `MemKind` to `Node`. It's OK to
do that here, because the code is not released yet.

Third, a new `FileDescriptor` type is introduced. It represents a
“generic” file descriptor, that a `VirtualFile` can hold. It contains
a value with the size of the larger size a file descriptor on any
platform, namely on Windows where a file descriptor is the size of a
pointer (compared to Unix where it's a `i32`). Then a bunch of
`TryInto` implementations are written to convert it to `RawFd` (on
Unix) or `RawHandle` (on Windows). I'm not confident it's the best
design for the moment, but it's better than before where Windows file
descriptors were impossible to represent. Note that we cannot make
`FileDescriptor` generic over the system for 2 reasons

1. We can activate multiple file systems at the same time, so the
   representation of a file descriptor can change between file system,

2. If we make `FileDescriptor` generic, then the size of `dyn
   VirtualFile` is unknown at compile-time, which makes impossible to
   use in `wasmer-wasi`, hence this design decision of taking the
   larger possible size for a file descriptor so that it fits all
   scenarios.