refactor(wasi): Move new store creation from ModuleCache to WasiRuntimeImplementation
Previously the ModuleCache had a ::new_store() method for creating a new
wasmer::Store.

A `trait Compiler` was introduced to make this extendable.

This is pretty pointless abstraction though and creates a lot of
complications for upstream users like Deploy.

It's much more sensible to move new_store() to the WasiRuntimeImplementation
trait, which was hereby done.

The `new_engine` method was also moved from the ModuleCache to
runtime/mod.rs for constructing default engines for the environment.