fix(wasi): Enables support for all Wasi clock types
Enables support for the wasi process and thread time clock types (2 and 3).

These were already implemented, but not correctly wired up due to a
duplication of the Clockid type.

This commit switches to using the correct Snapshot0Clockid everywhere.

In the future the "enum Clockid" should be removed, but that was not
done in this commit because currently the ./regenerate.sh script is
broken.