fix(wasi): Update fd stats after sync
Previously the fd_sync() WASI syscall implementation would not update
the FileStat statistics for a file descriptor upon flushing.

This leads to incorrect metadata on fd_fdstat_get syscalls, which breaks
some tests.

This commit changes the syscall implementation to update the stats.

NOTE: this is still a relatively suboptimal implementation, we should
probably always update the stats.

This will have to be done in the context of a general Inode system
rewrite.