Switch FileSystemCache to use checked artifact deserialization
We recently introduced safe artifact deserialization methods, which are a
much saner default than the unsafe variants, and only have small performance
overhead.

The CLI was switched to the the new, safe deserialization,
but the cache was not updated.

This commit and switches the FileSystemCache implementation to use the
checked deserialization variants.

NOTE: The Cache::load method should also be made safe, but that is a
breaking change.