chore(wasix): Remove deprecated thread_local functionality
thread_local support in wasix was previously implemented via custom
exports provided to modules.

This has been refactored to a more efficient (guest-internal)
implementation in both the wasix libc and the Rust fork.

We kept the exports and related functionality in for a while to allow for
modules to update, but now all the related functionality can safely be
removed.

Closes #3699