fix(virtual-fs): Fix horrible write performance for in-memory files
The write() implementation for in-memory files used by the TmpFs /
mem_fs was terrible: it used to re-allocate and copy the whole data if
the write position was not at the end of the file.

This re-writes the implementation to grow the buffer in place if
required.

Also adds some tests.

Closes #3903