fix: Prevent invalid file system merging errors
This commit changes the EmptyFileSystem read_dir() implementation to
always return an empty iterator when trying to read the root path "/".

This is done because some code paths expect the root to always be
readable.

This also makes sense theoretically, because an empty file system should
still be readable, just not contain any files.