Fix preopened current directory not working
This commit fixes a bug where creating a file with an implicit relative
path returns `0` indicating success, but no file is created on the host.
Here, implicit relative means a relative path that does not begin with
`./`. For example, calling `open("somefile", O_CREAT | O_WRONLY)` will
try to create a file called `somefile` in the current working directory.

The fix is to change the current directory if the user has explicitly
mounted any host directory to `.` in guest.  In this case, the user's
intention is clear: the current directory *is* the mounted host
directory.

fixes https://github.com/wasmerio/wasmer/issues/4361