feat(api): Add SharedMemoryHandle
Allows to trigger atomics related operations on a memory that supports
interacting with the atomics layer.

It adds a shared_handle() method to `Memory`.

Right now only final operations are exposed that wake up all atomic
waiters, and that allow disabling atomics.

This is required for forceful shutdown of instances that keep
hot-looping on atomics, and will be used from WASIX.