chore(wasix): Clean up WasiProcess state
Previously there was a redundant nested Arc and some general
shennanigans for how the state was shared and accessed.

This commit wraps the whole state in an Arc<> and removes the old nested
arc or the inner state.

It also makes a lot of fields private and exposes them through getters.

NOTE: the WasiProcessInner should also be made private