chore: Remove WAI codegen from wasmer-wasix-types
The WAI schema based code generation was used previously, but has not
been functioning for quite a while.
We have lots of manually defined types instead.

The codegen also required heavy patching to work in the first place.

Since this is a dead end, the automatic codegen is removed, since it is
dead code.