deps: Switch from unmaintained term_size to terminal_size
The term_size crate has been unmaintained for years.
The RUSTSEC warning recommends terminal-size instead.

The crate seems stable, and a review of the code shows it's safe to use
and does what is expected.
(no build.rs or similar, only calls to the relevant syscalls).

Gets rid of a RUSTSEC warning.