Relax fd_advise spec
`fd_advise` is spec'd so that the sum of offset and len cannot overflow.
This is unnecessary.  None of the other runtimes exhibit this behavior.
In addition, on Linux host, `posix_advise()` returns OK with `INT64_MAX`
as both offset and len.  This commit removes this part of the test.

fixes #4586