[skip ci] feat(wasmer_config): Make `Package` fields non-mandatory
This commit makes some fields of the `wasmer_config::package::Package`
type non mandatory. In particular, `name`, `version` and `description`
aren't mandatory anymore. This introduces *two ways* to describe unnamed
packages - no `package` section in the manifest and `package` section
with no `name` field in it, and a third kind of package in the package
taxonomy, which I deliberately name, here _unversioned packages_, that
is packages with a `name` (and any other field) but no `version`.