Allow wasmer-wasix to block the current thread
A WasiEnv now has a flag that signifies if it's allowed
to block the current thread and the flag is expected to
be referenced when the current thread can be blocked for
optimising the performance. In addition, the flag is set
true when wasm modules are invoked via the CLI. wasmer-wasix
uses a tokio runtime for many things according to #4299, but
tokio does not always bring the best performance in Wasmer
as described in #4738, hence this update.

Signed-off-by: Daichi Fukui <daichi1.fukui@toshiba.co.jp>