Allow more consistent sleep form of poll_oneoff
Allow different sleep form of poll_oneoff, which
just sleeps for relative time using std::thread::sleep.
That form of poll_oneoff is only available if there is
only one subscription and it is allowed to block
the current thread, contributing to more consistent
sleep times. This update would help improve
the real-time performance of wasmer.

Closes: #4738

Signed-off-by: Daichi Fukui <daichi1.fukui@toshiba.co.jp>