Fix fd with append flag can't seek correctly
This commit fixes an incompatible behavior (with other runtimes, and
POSIX) when an fd that was opened with `fdflags::append` cannot seek to
the right offset with `fd_seek`.  The append flag should not "lock" the
offset at the end, as it's possible to seek to another offset and read
from it.

Signed-off-by: Yage Hu <me@huyage.dev>