Fix append-ness of new fd affecting old fd
This commit fixes a case where opening a new file descriptor with the
append flag makes previously opened non-append fds append as well.

The fix is relatively simple, since the append-ness of an fd is already
tracked in the virtual fs implementation, there's no need to open the
actual file with the append flag.

This behavior is consistent across Linux and other Wasm runtimes.

fixes #5161

Signed-off-by: Yage Hu <me@huyage.dev>