feat(api): Add `id()` function to `wasmer::Engine`.
Adds a function to retrieve a unique identifier (backed by an
`AtomicU64`) for the top-level `Engine` type. The `id` is saved in the
`Engine` struct itself rather than types down the line. This fits better
with the current structure of the `sys` rt engines, which are defined in
the `wasmer_compiler` crate.