fix: pass readme content instead of readme file name
the readme field in wasmer.toml is supposed to be a pathbuf, but it is
being passed to the backend as a string. This should be the content
of the readme file, and not the file path itself.

the commit fixes this by reading the file, and passing the content
correctly.