make `AppConfig.name` optional
The wasmer-cli treats `AppConfig` objects without names as valid, but
this behavior is not reflected in the `AppConfig` schema.

This discrepancy caused issues when attempting to parse a "valid"
`AppConfig` using the wasmer-config create command, as the name field
was marked as required.

This commit makes the `AppConfig.name` field optional, thereby allowing
programs to successfully parse configurations without a name field.