fix(virtual-fs): OverlayFs: don't copy files that are never written
Previously te OverlayFs would unconditionally create copies of files on
shutdown if a file was opened with write flags, even if the file was
never written to.

This introduces redundant overhead and also causes unexpected behaviour for
downstream applications, because of unexpected changes to ctime and mtime.

Fixed by not initiating a copy in flush and shutdown for files that were
not written.