Simply object initialization path with relocations
Previous code relies on C glue code to piece metadata and actual
function addresses after linking together. This change refines the
process, so proper relocations are generated for each function /
trampoline when building the object. This way a linker should be able to
fulfill the task of filling in actual function addresses in the final
binary, there is no need to use glue code to do the memcpy work anymore.