riscv64: Implement indirect jump in gen_std_trampoline_riscv64()
This is the last step to have a functional trampoline for our use case.
After this point, we can focus in the code generation. At this moment,
we don't generate any code for the callee parameter of
wasmer_vm::trap::traphandlers::wasmer_call_trampoline, and it means that
we execute a 0, that is the encoding of unimp and we get a sigill that
allows us to verify that we jumped to the correct place.