riscv64: Remove the debug unimp's
While I was debugging I inserted these instructions to be sure
that the code generated by move_local() and move_location()
was not used before jumping to the trampoline, but it happened
that they were called at the beginning of the function generated
and it generated 4 unimp instructions, and since I was using
a x/4i command to inspect only the 4 initial instructions of the
generated code then I assumed that it was just garbage, but it
wasn't. The previous commit avoided the problem just jumping
over these instructions, but the correct solution was to
remove that debugging code (that was useless after we got a correct
trampoline) and remove the hack to jump over the instructions.