fix(tests): respect cargo --target argument correctly
When building the tests with --target cargo argument, we can't grab
the Cargo env. variables within the procedural macro as it is
something that's actually targeting the host (basically a compiler
plug-in). And so, we need to export the variables early in the top-level
build.rs file. Plus, for similar reasons, we can't use the `env!` macro
that maps to the variables at compile time.

Fixes: #5811