fix(cicd): Swap to new organizational ciuser token (#5902)
The integratoin test suite already supports fetching the most recent
artifact, so it'll run the tests using the newly built
"wasmer-cli-linux-x64".

The old behaviour here was to test against the dev environment. But
I'm swapping to instead run against production, which is much more
stable. This way we can see if there are any changes in the runtime
which will break any functionality exposed via the integration tests.