fix: Prevent filesystem merge for packages
Previously adding additional dependencies to a WASI environment would
lead to a recursive file system merge, which would clone all files in
memory, which is very expensive.

This was caused by a very unstructed creation and extension logic in the
WasiFs, which could take on a complicated form of <dyn Filesystem>.

There previosly were attempts to downcast and extend without merging,
but those would fail in common use cases.

This is fixed by adding a dedicated WasiFs::Overlay variant, which is a
dedicated representation for the common wasix package environment use
case. This allows to deterministically extend an existing overlay with
additional read-only mounts.